+++ app/routes/components/Toast/changelog.codegen.ts
@@ -1,3 +1,5 @@
 export const changelog = ''
   + '- 2025-09-20: Initial version.\n'
+  + '- 2025-09-23: Add support for pausing the auto-dismiss timer when hovering\n'
+  + '  over a toast & add basic styles.\n'
   ;