+++ src/components/Toast/Toast.tsx
@@ -4,16 +4,28 @@ import { createStore, type StoreApi } from "zustand";
 
 import styles from './Toast.module.scss';
 import { useTheme } from "../ThemeProvider";
+import { uniqueId } from "es-toolkit/compat";
+
+// Types
+// =============================================================================
+
+type Toast = {
+  id: string;
+  content: string;
+  timeoutHandler: ReturnType<typeof setTimeout> | null;
+}
 
 // ToastStore
 // =============================================================================
 
 type ToastStoreState = {
-  toasts: Array<string>;
+  toasts: Array<Toast>;
 };
 
 type ToastStoreActions = {
   addToast: (message: string) => void;
+  hoverToast?: (id: string) => void;
+  unhoverToast?: (id: string) => void;
 };
 
 type ToastStore = ToastStoreState & ToastStoreActions;
@@ -23,14 +35,47 @@ function buildToastStore() {
     toasts: [],
     
     addToast: (message) => {
-      setTimeout(() => {
+      const timeoutHandler = setTimeout(() => {
         set((state) => ({
-          toasts: state.toasts.slice(1),
+          toasts: state.toasts.filter((t) => t.id !== toast.id),
         }))
-      }, 3000);
+      }, 4000);
+
+      const id = uniqueId('toast-');
+      const toast = {
+        id,
+        content: message,
+        timeoutHandler,
+      };
 
       set((state) => ({
-        toasts: [...state.toasts, message],
+        toasts: [...state.toasts, toast],
+      }))
+    },
+    hoverToast: (id) => {
+      set((state) => ({
+        toasts: state.toasts.map((t) => {
+          if (t.id === id && t.timeoutHandler) {
+            clearTimeout(t.timeoutHandler);
+            return { ...t, timeoutHandler: null };
+          }
+          return t;
+        }),
+      }))
+    },
+    unhoverToast: (id) => {
+      set((state) => ({
+        toasts: state.toasts.map((t) => {
+          if (t.id === id && !t.timeoutHandler) {
+            const timeoutHandler = setTimeout(() => {
+              set((state) => ({
+                toasts: state.toasts.filter((toast) => toast.id !== id),
+              }))
+            }, 6000);
+            return { ...t, timeoutHandler };
+          }
+          return t;
+        }),
       }))
     },
   }));
@@ -69,6 +114,7 @@ const ToastContainer: FC = () => {
   const computedStyle = {
     '--toast-bg-color': 'white',
     '--toast-border-color': theme.colors.gray['300'],
+    '--toast-box-shadow': '0 2px 4px rgba(0, 0, 0, 0.05)',
   } as React.CSSProperties;
 
   return (
@@ -77,9 +123,14 @@ const ToastContainer: FC = () => {
       className={styles.toastContainer}
     >
       <ul className={styles.toastList}>
-        {toasts.map((toast, index) => (
-          <li key={index} className={styles.toast}>
-            {toast}
+        {toasts.map((toast) => (
+          <li
+            key={toast.id}
+            className={styles.toast}
+            onMouseEnter={() => store.getState().hoverToast?.(toast.id)}
+            onMouseLeave={() => store.getState().unhoverToast?.(toast.id)}
+          >
+            {toast.content}
           </li>
         ))}
       </ul>