+++ src/components/Toast/index.ts
@@ -1,3 +1,3 @@
-// Toast from copy-ui @ 2025-09-20
+// Toast from copy-ui @ 2025-09-23
 
 export { Toast, useToast } from "./Toast";