+++ src/components/Toast/CHANGELOG.md
@@ -1,1 +1,3 @@
 - 2025-09-20: Initial version.
+- 2025-09-23: Add support for pausing the auto-dismiss timer when hovering
+  over a toast & add basic styles.